requirements-production.txt
# Production-specific dependencies

# API enhancements
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6

# WebSocket support
websockets>=12.0
aioredis>=2.0.1

# Rate limiting and security
slowapi>=0.1.9
python-dotenv>=1.0.0

# Monitoring and observability
opentelemetry-api>=1.20.0
opentelemetry-sdk>=1.20.0
opentelemetry-instrumentation-fastapi>=0.41b0
opentelemetry-exporter-prometheus>=0.41b0
prometheus-client>=0.18.0

# Tracing
langfuse>=2.0.0

# Database and caching
redis>=5.0.0
aioredis>=2.0.1

# Guardrails and safety
presidio-analyzer>=2.2.0
presidio-anonymizer>=2.2.0

# Export and reporting
reportlab>=4.0.0
markdown>=3.5.0

# Logging
python-json-logger>=2.0.7
